Naya/NyayAssist/requirements.txt
# ============================================
# NyayAssist Backend Requirements
# ============================================

# ----------------------
# Core Web Framework
# ----------------------
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.18
pydantic[email]>=2.10.0

# ----------------------
# LLM & AI
# ----------------------
google-generativeai>=0.8.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
langchain-huggingface>=0.1.0
langchain-google-genai>=2.0.0

# ----------------------
# Vector Store & Embeddings
# ----------------------
faiss-cpu>=1.9.0
sentence-transformers>=3.3.0

# ----------------------
# PDF Processing
# ----------------------
PyPDF2>=3.0.0

# ----------------------
# Database (MySQL)
# ----------------------
pymysql>=1.1.0
sqlalchemy>=2.0.0
aiomysql>=0.2.0
cryptography>=43.0.0

# ----------------------
# Security & Auth
# ----------------------
bcrypt>=4.2.0
passlib[bcrypt]>=1.7.4

# ----------------------
# Utilities
# ----------------------
python-dotenv>=1.0.0
requests>=2.32.0

# ----------------------
# Optional: LlamaIndex (if using)
# ----------------------
# llama-index>=0.12.0
# llama-index-llms-gemini>=0.4.0

# ----------------------
# Optional: Streamlit App
# ----------------------
streamlit>=1.40.0